cybot-gui/requirements.txt
pyserial
pygame
math
